added flow control to ensure a user can only delete a quesiton if they are signed in and it is a question they created

+++ app/controllers/questions_controller.rb
@@ -1,5 +1,3 @@
-require 'pry'
-
 class QuestionsController < ApplicationController
   def index
     @questions = Question.all
@@ -18,15 +16,14 @@ class QuestionsController < ApplicationController
   end
 
   def create
-    if session[:user_id] == nil
+    if !current_user
       flash[:notice] = "You must be signed in to create a new question"
       render 'new'
     else
-      user_id = session[:user_id]
       title = question_params[:title]
       description = question_params[:description]
 
-      @question = Question.create(user_id: user_id, title: title, description: description)
+      @question = Question.create(user_id: current_user.id, title: title, description: description)
 
       if @question.save
         flash[:notice] = "You Successfully Created Your New Question!"
@@ -40,10 +37,10 @@ class QuestionsController < ApplicationController
   def update
     @question = Question.find(params[:id])
 
-    if session[:user_id] == nil
+    if !current_user
       flash[:notice] = "You must be signed in to edit a question"
       redirect_to edit_question_path(params[:id])
-    elsif session[:user_id] != @question.user_id
+    elsif current_user.id != @question.user_id
       flash[:notice] = "You can only edit your own questions."
       redirect_to edit_question_path(params[:id])
     else
@@ -57,11 +54,21 @@ class QuestionsController < ApplicationController
 
   def destroy
     @question = Question.find(params[:id])
-    @question.destroy
-    answers = Answer.where(question_id: @question.id)
-    answers.delete_all
 
-    redirect_to questions_path
+    if !current_user
+      flash[:notice] = "You must be signed in to delete a question."
+      redirect_to @question
+    elsif current_user.id != @question.user_id
+      flash[:notice] = "You can only delete a question you've created."
+      redirect_to @question
+    else
+      @question.destroy
+      answers = Answer.where(question_id: @question.id)
+      answers.delete_all
+      flash[:notice] = "Your question has been deleted."
+
+      redirect_to questions_path
+    end
   end
 
   private
